refactor e sistemazione test

+++ TddCodeCata/src/tdd1/supermarket/ArticleFactory.java
@@ -4,7 +4,7 @@ public class ArticleFactory
 {
 	
 	
-	public Article article(String name, double price) {
+	public Article article(String name, int price) {
 		
 		return new Article(name, new Price(price));
 	}
